app/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
shapely
jinja2
pydantic
python-dotenv
python-multipart
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
sounddevice
edge-tts
pygame
faster-whisper